requirements.txt
PyGithub==2.8.1
pandas==2.3.3
python-dotenv==1.2.1
openai==2.7.2
pypdf==6.2.0
gradio==5.49.1
fastapi==0.121.1
uvicorn==0.38.0